additional work on UserController and EventsController to find resource by ID, still not working as intended

+++ volapp/src/main/java/com/volapp/events/Events.java
@@ -28,7 +28,7 @@ public class Events{
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     
     
     public Events() {}
@@ -46,7 +46,7 @@ public class Events{
     	this.eventDescription = events.eventDescription;
     }
     
-    public Events (Long id, String charityName, String eventName, String eventStreet, String eventCity, String eventState, String eventZip, String eventDate, String eventTime, String eventDescription) {
+    public Events (Integer id, String charityName, String eventName, String eventStreet, String eventCity, String eventState, String eventZip, String eventDate, String eventTime, String eventDescription) {
     	this.id = id;
     	this.charityName = charityName;
     	this.eventName = eventName;
@@ -61,12 +61,12 @@ public class Events{
     
 
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
